locate blocks
